pyproject.toml
[project]
name = "igh-airflow"
version = "0.1.0"
description = "Apache Airflow orchestration for the IGH Data Pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.1.0",
    "apache-airflow-providers-fab>=2.0.0",
    "apache-airflow-providers-celery>=3.0.0",
    "redis>=5.0.0,<6.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "igh-data-sync",
    "igh-data-transform",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dags", "config"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.uv.sources]
igh-data-sync = { git = "https://github.com/akvo/igh-data-sync.git" }
igh-data-transform = { git = "https://github.com/akvo/igh-data-transform.git", branch = "silver-to-gold-transformation" }
